+++ src/js-framework/polyfill/index.js
@@ -1,4 +1,4 @@
 import './objectAssign'
-import './promise'
 import './setTimeout'
 import './consolelog'
+import './promise'