+++ src/js-framework/CHANGELOG.md
@@ -3,5 +3,7 @@
 ## v0.13.10
 
 - add `updateFinish` action after `click/fireEvent`
+- add `console.log/debug/info/warn/error`
+- update test for coverage
 - add a testcase for `track-by`
 - fixed #323. fix computed in repeat scope